rename support/render/healthjson => support/render/health
