+++ support/render/health/doc.go
@@ -1,0 +1,6 @@
+// Package health contains simple utilities for implementing a /health endpoint
+// that adheres to the requirements defined in the draft IETF network working
+// group standard, Health Check Response Format for HTTP APIs.
+//
+// https://tools.ietf.org/id/draft-inadarei-api-health-check-01.html
+package health
