+++ support/render/health/handler.go
@@ -1,4 +1,4 @@
-package healthjson
+package health
 
 import (
 	"net/http"
